Add tests for custom script sort
